primeira parte do exercicio sobre heranca

+++ projeto-java/src/exercicios/POO/terceiro/Pessoa.java
@@ -4,8 +4,8 @@ public class Pessoa {
     String nome;
     int idade;
     String ID;
-    float peso;
-    float altura;
+    double peso;
+    double altura;
 
     public Pessoa() {
         this.nome = nome;
@@ -21,7 +21,7 @@ public class Pessoa {
         this.ID = ID;
     }
 
-    public Pessoa(String nome, int idade, String ID, float peso, float altura) {
+    public Pessoa(String nome, int idade, String ID, double peso, double altura) {
         this.nome = nome;
         this.idade = idade;
         this.ID = ID;
@@ -32,7 +32,7 @@ public class Pessoa {
     public static void main (String args[]){
         Pessoa pessoaConstrutorVazio = new Pessoa();
         Pessoa pessoaConstrutorMenosParametros = new Pessoa("Hugo",28,"!@#dsfd121312");
-        Pessoa pessoaConstrutorCompleto = new Pessoa("Hugo",28,"!@#dsfd121312", 83.5,1.84);
+        Pessoa pessoaConstrutorCompleto = new Pessoa("Hugo",28,"!@#dsfd121312",83.5,1.84);
         //Pessoa pessoaConstrutorIncorreto = new Pessoa("Hugo",28);
         //Incorreto pois o construtor com 2 parametros nao está criado
         System.out.println("Hello World");
